final commit

+++ src/main/java/sk/vaii/sem/semestralna_praca_vaii_backend/part_film/service/GenreService.java
@@ -1,6 +1,7 @@
 package sk.vaii.sem.semestralna_praca_vaii_backend.part_film.service;
 
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import sk.vaii.sem.semestralna_praca_vaii_backend.part_film.entity.Genre;
 import sk.vaii.sem.semestralna_praca_vaii_backend.part_film.repository.GenreRepository;
@@ -13,12 +14,15 @@ import java.util.Optional;
 public class GenreService {
     private final GenreRepository genreRepository;
 
-    public Genre addGenre(Genre genre) {
+    public Genre addGenre(String genreName) {
+        Genre genre = new Genre();
+        genre.setName(genreName);
+
         return this.genreRepository.save(genre);
     }
 
     public List<Genre> getAllGenres() {
-        return this.genreRepository.findAll();
+        return this.genreRepository.findAll(Sort.by(Sort.Direction.ASC, "id"));
     }
 
     public Optional<Genre> getGenreById(Long id) {
@@ -29,7 +33,16 @@ public class GenreService {
         return this.genreRepository.findByName(name);
     }
 
+    public List<Genre> getGenresOfFilm(Long filmId) {
+        return this.genreRepository.getGenresOfFilm(filmId);
+    }
+
     public void deleteGenre(Long id) {
-        this.genreRepository.deleteById(id);
+        Genre genre = this.genreRepository.getById(id);
+        if (genre.getFilms().isEmpty()) {
+            this.genreRepository.deleteById(id);
+        } else {
+            throw new IllegalStateException("Genre can't be assigned to any movie.");
+        }
     }
 }
